Refactor: Make CommonQuestionService
